dynamic content length: mbedtls_ssl_confirm_content_len
This patch introduces mbedtls_ssl_confirm_content_len(), which
ensures that either the in or out record buffer is at least as
large enough for the given size.  If necessary, the buffer is
reallocated.

Length checks are preceded by calls to mbedtls_ssl_confirm_content_len()
attempting to resize the buffer being checked such that the check
will always pass.

mbedtls_ssl_confirm_content_len() will limit the allocation to 16384 +
the overhead constant if asked to allocate anything larger than 16384.

Initially, the buffer size is zero.

The in and out buffer sizes are handled separately.

The effect of this is to make the buffers dynamically "stretchy", such
that initially too small buffers simply adapt instead of fail out as
before.

For the test client, instead of a buffer allocation of 32KB + the overhead,
the client side only requires an allocation of 6625 bytes (1933 out and
4692 in).

Signed-off-by: Andy Green <andy@warmcat.com>